+++ internal/query/execute_query.go
@@ -143,7 +143,7 @@ func readAll(ctx context.Context, r *streamResult) error {
 	}
 }
 
-func readResultSet(ctx context.Context, r *streamResult) (_ *resultSet, finalErr error) {
+func readResultSet(ctx context.Context, r *streamResult) (_ *resultSetWithClose, finalErr error) {
 	defer func() {
 		_ = r.Close(ctx)
 	}()
@@ -161,7 +161,10 @@ func readResultSet(ctx context.Context, r *streamResult) (_ *resultSet, finalErr
 		return nil, xerrors.WithStackTrace(err)
 	}
 
-	return rs, nil
+	return &resultSetWithClose{
+		resultSet: rs,
+		close:     r.Close,
+	}, nil
 }
 
 func readMaterializedResultSet(ctx context.Context, r *streamResult) (_ *materializedResultSet, finalErr error) {